+++ packages/framework/src/Model/Category/CategoryData.php
@@ -57,6 +57,11 @@ class CategoryData
      */
     public $pluginData;
 
+    /**
+     * @var string|null
+     */
+    public $uuid;
+
     public function __construct()
     {
         $this->name = [];