+++ packages/frontend-api/src/ShopsysFrontendApiBundle.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shopsys\FrontendApiBundle;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+class ShopsysFrontendApiBundle extends Bundle
+{
+}